+++ react/components/Shelf/types.d.ts
@@ -1,4 +1,5 @@
 type NewtailMediaShelfProps = {
+  active?: boolean
   placementName?: string
   placementNameAdmin?: string
   quantity?: number